lerna create @firestore-simple/admin
