+++ packages/admin/README.md
@@ -1,0 +1,11 @@
+# `@firestore-simple/admin`
+
+> TODO: description
+
+## Usage
+
+```
+const admin = require('@firestore-simple/admin');
+
+// TODO: DEMONSTRATE API
+```
