+++ packages/admin/__tests__/query_on_snapshot.test.ts
@@ -1,4 +1,4 @@
-import { FirestoreSimpleAdmin } from '../../src'
+import { FirestoreSimpleAdmin } from '../src'
 import { createRandomCollectionName, deleteCollection, initFirestore } from './util'
 
 // Workaround for flaky onSnapshot callback tests.
@@ -8,11 +8,13 @@ interface Book {
   id: string,
   bookTitle: string,
   created: Date,
+  bookId: number,
 }
 
 interface BookDoc {
   book_title: string,
   created: Date,
+  book_id: number,
 }
 
 const firestore = initFirestore()
@@ -20,13 +22,14 @@ const collectionPath = createRandomCollectionName()
 const firestoreSimple = new FirestoreSimpleAdmin(firestore)
 
 // Skip reason: Sometimes real Firestore is unstable so it will be replaced emulator test.
-describe.skip('on_snapshot test', () => {
+describe.skip('query on_snapshot test', () => {
   const dao = firestoreSimple.collection<Book, BookDoc>({
     path: collectionPath,
     encode: (book) => {
       return {
         book_title: book.bookTitle,
         created: book.created,
+        book_id: book.bookId,
       }
     },
     decode: (doc) => {
@@ -34,6 +37,7 @@ describe.skip('on_snapshot test', () => {
         id: doc.id,
         bookTitle: doc.book_title,
         created: doc.created.toDate(), // Firestore timestamp to JS Date
+        bookId: doc.book_id,
       }
     },
   })
@@ -43,6 +47,7 @@ describe.skip('on_snapshot test', () => {
     const addedDoc = {
       bookTitle: 'exists',
       created: new Date(),
+      bookId: 1,
     }
     const addedId = await dao.add(addedDoc)
     existsDoc = {
@@ -57,24 +62,26 @@ describe.skip('on_snapshot test', () => {
 
   it('observe add change', async () => {
     const doc = {
-      bookTitle: 'add',
+      bookTitle: 'query_add',
       created: new Date(),
+      bookId: 2
     }
 
     const promise = new Promise((resolve) => {
-      dao.onSnapshot((querySnapshot, toObject) => {
-        querySnapshot.docChanges().forEach((change) => {
-          if (change.type === 'added' && change.doc.data().book_title === doc.bookTitle) {
-            const changedDoc = toObject(change.doc)
+      dao.where('book_id', '==', doc.bookId)
+        .onSnapshot((querySnapshot, toObject) => {
+          querySnapshot.docChanges().forEach((change) => {
+            if (change.type === 'added') {
+              const changedDoc = toObject(change.doc)
 
-            expect(changedDoc).toEqual({
-              ...doc,
-              id: expect.anything()
-            })
-            resolve()
-          }
+              expect(changedDoc).toEqual({
+                ...doc,
+                id: expect.anything(),
+              })
+              resolve()
+            }
+          })
         })
-      })
     })
 
     await new Promise((resolve) => setTimeout(resolve, 100)) // for async stability
@@ -84,21 +91,24 @@ describe.skip('on_snapshot test', () => {
 
   it('observe set changes', async () => {
     const doc = {
-      ...existsDoc,
-      bookTitle: 'set'
+      ...existsDoc!,
+      bookTitle: 'query_set',
     }
 
     const promise = new Promise((resolve) => {
-      dao.onSnapshot((querySnapshot, toObject) => {
-        querySnapshot.docChanges().forEach((change) => {
-          if (change.type === 'modified') {
-            const changedDoc = toObject(change.doc)
+      // where('book_title', '==', existsDoc) is not triggered modify existsDoc.
+      // I don't know why, so book_id is hack for resolve this issue.
+      dao.where('book_id', '==', 1)
+        .onSnapshot((querySnapshot, toObject) => {
+          querySnapshot.docChanges().forEach((change) => {
+            if (change.type === 'modified') {
+              const changedDoc = toObject(change.doc)
 
-            expect(changedDoc).toEqual(doc)
-            resolve()
-          }
+              expect(changedDoc).toEqual(doc)
+              resolve()
+            }
+          })
         })
-      })
     })
 
     await new Promise((resolve) => setTimeout(resolve, 100)) // for async stability
@@ -107,25 +117,13 @@ describe.skip('on_snapshot test', () => {
   })
 
   it('observe delete change', async () => {
-    // prepare specific doc for delete onSnapshot()
-    // because onSnapshot() also triggerd deleteCollection() events and it will be confilict.
-    const deletedDoc = {
-      bookTitle: 'deleted',
-      created: new Date(),
-    }
-    const deletedId = await dao.add(deletedDoc)
-
     const promise = new Promise((resolve) => {
-      dao.onSnapshot((querySnapshot, toObject) => {
+      dao.where('book_title', '==', existsDoc.bookTitle).onSnapshot((querySnapshot, toObject) => {
         querySnapshot.docChanges().forEach((change) => {
-          if (change.type === 'removed' && change.doc.data().book_title === deletedDoc.bookTitle) {
+          if (change.type === 'removed' && change.doc.data().book_title === existsDoc.bookTitle) {
             const changedDoc = toObject(change.doc)
 
-            expect(changedDoc).toEqual({
-              id: expect.anything(),
-              bookTitle: deletedDoc.bookTitle,
-              created: deletedDoc.created,
-            })
+            expect(changedDoc).toEqual(existsDoc)
             resolve()
           }
         })
@@ -133,7 +131,7 @@ describe.skip('on_snapshot test', () => {
     })
 
     await new Promise((resolve) => setTimeout(resolve, 100)) // for async stability
-    await dao.delete(deletedId)
+    await dao.delete(existsDoc.id)
     await promise
   })
 })
