+++ packages/admin/src/index.ts
@@ -1,0 +1,8 @@
+// Ailias name for campatible before v6.0.0
+export { FirestoreSimpleAdmin as FirestoreSimple } from './admin'
+export type { AdminEncodable as Encodable, AdminDecodable as Decodable } from './types'
+
+// Admin SDK
+export { FirestoreSimpleAdmin } from './admin'
+export type { AdminCollection } from './collection'
+export type { AdminEncodable, AdminDecodable } from './types'
